password validation complete
now when the user types in the password, it is being checked against the regex pattern.
if the user clicks away from the password the validation message will disappear.
speaking of which, the validation message is now in a window of its own and is not just a text message.
the validation message will now appear in red if the password is invalid and black if the password is valid.